% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_formats_dataframe.R
\name{change_formats_dataframe}
\alias{change_formats_dataframe}
\title{Change formats of columns in dataframe}
\usage{
change_formats_dataframe(df, formats_df = formats, char_vars = FALSE,
  factor_vars = FALSE, post_dm = FALSE, var_selection = c("_all"))
}
\arguments{
\item{df}{dataframe}

\item{formats_df}{dataframe defining the formats. Must be in standard format containing columns \code{Variable_name}, \code{Import_format}
and \code{Sorting_order}.}

\item{char_vars}{Logical if \code{TRUE}, all \code{variables} that are
defined as \code{"String"} in column \code{Variable_type} of the formats file will be
changed to \code{as.character}.}

\item{factor_vars}{Logical if \code{TRUE}, all \code{variables} that are
defined as \code{"Labelled num"} in column \code{Variable_type} and
\code{"Ordinal"} or \code{"Nominal"} in column \code{Measurement_level} of the formats file will be
changed to \code{sjlabelled::as_label, keep.labels = TRUE}.}

\item{post_dm}{Logical; if \code{FALSE}, all \code{variables} are added as part of data management steps
defined as \code{"not imported"} in column \code{Import_format} of the formats file will be
omitted.}

\item{var_selection}{Character vector of selected variables will be changed. Default \code{var_selection = c("all")} means that
all variables are selected.}
}
\value{
dataframe
}
\description{
Apply column formats as defined in formats file to dataframe.
}
