% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_unused_factor_levels.R
\name{drop_unused_factor_levels}
\alias{drop_unused_factor_levels}
\title{Drop unused factor levels for dataframe}
\usage{
drop_unused_factor_levels(df, var_selection = c("_all"))
}
\arguments{
\item{df}{dataframe}

\item{var_selection}{Character vector of selected variables to be set missing. Default \code{var_selection = c("all")} means that
all variables are selected}
}
\value{
dataframe
}
\description{
Drop unused factor levels from all factor variables in the dataframe.
}
