% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_vals_dataframe.R,
%   R/label_vars_dataframe.R
\name{label_vars_dataframe}
\alias{label_vars_dataframe}
\title{Apply value labels to dataframe}
\usage{
label_vars_dataframe(df, formats_df = formats, post_dm = FALSE)

label_vars_dataframe(df, formats_df = formats, post_dm = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{formats_df}{dataframe defining the formats. Must be in standard format containing columns \code{Variable_name}, \code{Value_Label_1-x}.
If needed run function \code{read_formats_xlsx.R} beforehand.}

\item{post_dm}{Logical; if \code{FALSE}, all \code{variables} are added as part of data management steps
defined as \code{"not imported"} in column \code{Import_format} of the formats file will be
omitted.}

\item{df}{dataframe}

\item{formats_df}{dataframe defining the formats. Must be in standard format containing columns \code{Variable_name}, \code{Import_format}
and \code{Sorting_order}.}

\item{post_dm}{Logical; if \code{FALSE}, all \code{variables} are added as part of data management steps
defined as \code{"not imported"} in column \code{Import_format} of the formats file will be
omitted.}
}
\value{
dataframe

dataframe
}
\description{
Apply value labels as defined in formats file to all variables in the dataframe.

Apply variable labels as defined in formats file to dataframe.
}
