% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msAutolabelR.R
\docType{package}
\name{msAutolabelR}
\alias{msAutolabelR}
\title{msAutolabelR}
\description{
Helper Functions to Automatically Label Dataframes Using a Format Table
}
\section{Main functions}{


Functions include label_values_dataframe, label_vars_dataframe, read_formats, set_missing_values.
}

\section{Package context}{

This package follows the principles of the "tidyverse".
}

\keyword{internal}
