% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formats_xlsx.R
\name{read_formats_xlsx}
\alias{read_formats_xlsx}
\title{Import Formats from .xlsx}
\usage{
read_formats_xlsx(infile_formats, sheet = NULL)
}
\arguments{
\item{infile_formats}{file}

\item{sheet}{string}
}
\value{
dataframe
}
\description{
Import formats from a predefined xlsx file.
}
