% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_dataframe.R
\name{sort_dataframe}
\alias{sort_dataframe}
\title{Sort Dataframe}
\usage{
sort_dataframe(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
Sort an existing dataset according to the order of variables defined in the format table.
}
\examples{
sort_dataframe(iris)
}
