% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_values_dataframe.R
\name{label_values_dataframe}
\alias{label_values_dataframe}
\title{Apply value labels to dataframe}
\usage{
label_values_dataframe(df, formats_df = formats, post_dm = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{formats_df}{dataframe defining the formats. Must be in standard format containing columns \code{Variable_name}, \code{Value_Label_1-x}.
If needed run function \code{read_formats_xlsx.R} beforehand.}

\item{post_dm}{Logical; if \code{FALSE}, all \code{variables} are added as part of data management steps
defined as \code{"not imported"} in column \code{Import_format} of the formats file will be
omitted.}
}
\value{
dataframe
}
\description{
Apply value labels as defined in formats file to all variables in the dataframe.
}
